server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
openenv-core>=0.2.2
rich>=13.0.0
